Function to provide a quick table for ANCOVA analyses with linear models

This function takes a continuous outcome (specified by var) and a binary covariate (likely a treatment assignment variable specified by arm), and fit a linear regression adjusted for additional covariates (specified by covariates, which is a vector of string values. The function will output the observed means and standard deviations stratified by arm, as well as LS means and standard errors. The p-value reported will be the overall ANOVA p-value of the term specified in arm.

Usage

qtab_lm_anova(df, var, arm, covariates = NULL, ...)

Arguments

df

tibble/DataFrame The data frame of interest

var

string Name of the outcome variable

arm

string Name of the primary treatment variable (or any other factor variable) that is of primary interest

covariates

vector. A vector of covariates for adjustment

...

Additional arguments passed to lm function

Value

A flextable table

Examples

data("rand_adlb")
proc_df <- rand_adlb |> 
    dplyr::rename_all(tolower) |> 
    dplyr::filter(paramcd == "ALT") |> 
    dplyr::filter(avisit == "WEEK 5 DAY 36")
qtab_lm_anova(df = proc_df, var = "aval", 
    arm = "trt01a", covariates = NULL)

Treatment

N

Mean (SD)

LS Means (SE)

p-value

A: Drug X

135

17.77 (9.32)

17.77 (0.82)

0.1714

B: Placebo

141

19.22 (9.73)

19.22 (0.8)

C: Combination

124

19.94 (9.49)

19.94 (0.85)